requirements.txt
fake_useragent
beautifulsoup4
requests
pypandoc
selenium
Pillow
customtkinter
